[service/tx] rename txConfirmLogic.dart to submitToJS.dart
