+++ lib/service/tx/submitTxWrappers.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:encointer_wallet/common/components/passwordInputDialog.dart';
 import 'package:encointer_wallet/service/substrate_api/api.dart';
-import 'package:encointer_wallet/service/tx/txConfirmLogic.dart';
+import 'package:encointer_wallet/service/tx/submitToJS.dart';
 import 'package:encointer_wallet/store/app.dart';
 import 'package:encointer_wallet/store/encointer/types/communities.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
